bot/requirements.txt
uvicorn
fastapi
psycopg2-binary
sqlalchemy
python-multipart
sshpubkeys
python-openstackclient
requests